+++ project.yml
@@ -1,0 +1,22 @@
+toolchain_path: "/Library/Developer/Toolchains/swift-latest/usr/lib/swift/macosx/"
+
+name: SwiftTensorflow
+options:
+  bundleIdPrefix: com.swiftTensorflow
+
+targets:
+  SwiftTensorflow:
+    type: tool
+    platform: macOS
+    deploymentTarget: "10.13"
+    sources: 
+        - Sources
+    dependencies:
+      - framework: /Library/Developer/Toolchains/swift-latest/usr/lib/swift/macosx/libtensorflow_framework.so
+      - framework: /Library/Developer/Toolchains/swift-latest/usr/lib/swift/macosx//libtensorflow.so
+      - framework: libpython2.7.tbd
+    settings:
+      SWIFT_OPTIMIZATION_LEVEL: -O
+      LD_RUNPATH_SEARCH_PATHS: /Library/Developer/Toolchains/swift-latest/usr/lib/swift/macosx/
+      OTHER_LDFLAGS:  -lpython
+
